CourseService
Atualização para correção erro 500 - não deixar deletar um curso qdo o mesmo esta associado a uma alocação

+++ src/main/java/com/project/professor/allocation/service/CourseService.java
@@ -7,6 +7,7 @@ import org.hibernate.service.spi.ServiceException;
 import org.springframework.stereotype.Service;
 
 import com.project.professor.allocation.entity.Course;
+import com.project.professor.allocation.repository.AllocationRepository;
 import com.project.professor.allocation.repository.CourseRepository;
 import com.project.professor.allocation.service.exception.EntityNotFoundException;
 
@@ -16,10 +17,12 @@ import net.bytebuddy.implementation.bytecode.Throw;
 public class CourseService {
 
 	private final CourseRepository courseRepository;
+	private final AllocationRepository allocationRepository;
 
-	public CourseService(CourseRepository courseRepository) {
+	public CourseService(CourseRepository courseRepository, AllocationRepository allocationRepository) {
 		super();
 		this.courseRepository = courseRepository;
+		this.allocationRepository = allocationRepository;
 	}
 
 	public List<Course> findByNameContaining(String name) {
@@ -53,7 +56,7 @@ public class CourseService {
 	}
 
 	public void deleteById(Long id) throws EntityNotFoundException {
-		if (id != null && courseRepository.existsById(id)) {
+		if (id != null && courseRepository.existsById(id)&& allocationRepository.findByCourseId(id) != null) {
 			courseRepository.deleteById(id);
 		} else {
 			throw new EntityNotFoundException("Course doesn't find");
